+++ Action/CollectionLightEffect.cpp
@@ -15,9 +15,9 @@ CollectionLightEffect::CollectionLightEffect(GameObject* _owner, const Vector3& 
 	SetPosition(position);
 
 	// ランダムな速度を得る
-	speed = (rand() % 400 + 200);
+	speed = (float)(rand() % 400 + 200);
 	// ランダムな初期の値を得る
-	value = (rand() % 180);
+	value = (float)(rand() % 180);
 	sinV = Math::Sin(value);
 
 	particleComponent->SetScale(scale);
